web/unbreakable-crypto/app/requirements.txt
Flask==1.1.2
Flask_APScheduler==1.12.2
requests==2.25.1
Pillow==9.0.0
pycryptodome==3.12.0
pycryptodomex==3.12.0
pyppeteer==0.2.6
